+++ lib/__tests__/autocomplete-test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
 
-import Autocomplete from '../autocomplete-temp';
+import Autocomplete from '../autocomplete';
 import {
   getStates,
   getCategorizedStates,
